requirements.txt
# pip install -r requirements.txt

Flask
sentence-transformers
pandas
numpy
faiss-gpu